public_sentiment_pipeline/requirements.txt
pytest
pytest-cov
pylint
requests
psycopg2-binary
python-dotenv
pandas
boto3
pytz
nltk